+++ package.json
@@ -116,7 +116,7 @@
     "rimraf": "^3.0.2",
     "serve": "^12.0.1",
     "sharp": "^0.29.1",
-    "stylelint": "^13.10.0",
+    "stylelint": "^14.2.0",
     "tslib": "^2.3.1",
     "typescript": "^4.5.2"
   },
