+++ packages/docusaurus-theme-translations/update.d.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import type {TranslationFileContent} from '@docusaurus/types';
+
+export function extractThemeCodeMessages(
+  targetDirs?: string[],
+): Promise<TranslationFileContent>;
