+++ packages/stylelint-copyright/__tests__/index.test.js
